Move updateItemQuantity to merlins connector
