+++ web/app/containers/cart/actions.js
@@ -10,8 +10,7 @@ import {
     CART_WISHLIST_MODAL,
     ADD_TO_WISHLIST_URL
 } from './constants'
-// import {removeFromCart} from '../../store/cart/actions'
-import {removeFromCart} from '../../integration-manager/cart/commands'
+import {removeFromCart, updateItemQuantity} from '../../integration-manager/cart/commands'
 import {makeFormEncodedRequest} from 'progressive-web-sdk/dist/utils/fetch-utils'
 import {getUenc} from '../product-details/selectors'
 import {addNotification} from '../app/actions'
@@ -85,6 +84,17 @@ export const openRemoveItemModal = (itemId) => {
 
 export const removeItem = (itemID) => (dispatch) => {
     return dispatch(removeFromCart(itemID))
+        .catch((error) => {
+            dispatch(addNotification({
+                content: error.message,
+                id: 'cartUpdateError',
+                showRemoveButton: true
+            }))
+        })
+}
+
+export const updateItem = (itemId, itemQuantity) => (dispatch) => {
+    return dispatch(updateItemQuantity(itemId, itemQuantity))
         .catch((error) => {
             dispatch(addNotification({
                 content: error.message,
