+++ web/app/integration-manager/cart/commands.js
@@ -9,3 +9,5 @@ export const getCart = (...args) => connector.getCart(...args)
 export const fetchCartPageData = (...args) => connector.fetchCartPageData(...args)
 
 export const removeFromCart = (...args) => connector.removeFromCart(...args)
+
+export const updateItemQuantity = (...args) => connector.updateItemQuantity(...args)
